main1/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
Pillow==4.3.0

# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
# Flask==1.1.2

# Flask Packages
# Flask-SQLAlchemy==2.4.4
# SQLAlchemy==1.3.20
# Flask-Script==2.0.6
# Flask-Cors==3.0.9
# requests==2.25.0
# mysqlclient==2.0.1
# pika==1.1.0
